Add files via upload
This utility will convert audio to text, convert pdf to text, convert text to pdf, convert mp3 to flac, convert flac to wav, convert mp3 to wav. It does all of this to convert multiple text or pdfs in a directory to an openai embedding parquet. I want to thank!!! Youtubers : Data Independent, Shweta Lodha, Kamalraj M M and OPENAI!!!!